backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary